Rename composables files
